Fixed structure and component naming. (#12)
