+++ src/components/SummaryCard.tsx
@@ -1,0 +1,5 @@
+export function SummaryCard() {
+    return (
+        <div>Summary Card</div>
+    )
+}
