+++ src/pages/Dashboard.tsx
@@ -1,12 +1,14 @@
 import { Header } from '../components/Header'
-import { DashboardMain } from '../components/DashboardMain'
+import { WelcomeBack } from '../components/WelcomeBack'
+import { SummaryCard } from '../components/SummaryCard'
 import './Dashboard.css'
 
 export function Dashboard() {
     return (
-        <div className='dashboard-container flex flex-col min-sm:ml-[16rem] px-5 py-5 gap-8'>
+        <main className='dashboard-container flex flex-col min-sm:ml-[16rem] px-5 py-5 gap-8'>
             <Header />
-            <DashboardMain />
-        </div>
+            <WelcomeBack />
+            <SummaryCard />
+        </main>
     )
 }
